tugas 06 - 23

+++ application/views/blog.php
@@ -48,40 +48,58 @@
       <li>
       <a href="<?php echo site_url('about/'); ?>">About</a>
       </li>
-      <li>
-      <a href="<?php echo site_url('blog/'); ?>">Blog</a>
-      </li>
+            <li>
+              <a href="<?php echo site_url('blog/'); ?>">Artikel</a>
+            </li>
+            <li>
+              <a href="<?php echo site_url('kategori/'); ?>">Kategori</a>
+            </li>
           </ul>
         </div><!--/.nav-collapse -->
       </div>
     </div>
   <!-- +++++ Second Post +++++ -->
   <div id="white">
+    <?php if( !empty($artikel) ) : ?> 
       <div class="container">
       <a href="<?php echo site_url('blog/tambah/'); ?>" class="btn btn-success">Tambah</a>
       <div class="row">
+        <?php foreach ($artikel as $key): ?>
         <div class="col-lg-8 col-lg-offset-2">
 
 
 
-<?php foreach ($artikel as $key): ?>
+
         <div class="col-xs-4 col-sm-4 col-md-4 col-lg-4">
           <table style="margin-bottom: 30px;">
             <tr>
               <td>
-                <a href="<?php echo base_url(). 'blog/detail/'. $key->id ?>" style="color: black;">
+                
+
+                   <?php if( $key->image ) : ?>
                   <img src="<?php echo base_url(). 'img/'. $key->image ?>" alt="Image" width="500" height="400">
+                  <?php ; else : ?>
+                            <img class="card-img-top" data-src="holder.js/100px190?theme=thumb&bg=eaeaea&fg=aaa&text=Thumbnail" alt="Card image cap">
+                            <?php endif; ?>
+                                   
+
+                                        <br>
+                            <small class="text-success text-uppercase"><?php echo $key->cat_name ?></small>
+                            <br>
+
+
                   <br>
-                  <center><h1><?php echo $key->judul ?></h1></center>
+                  <center><h1><?php echo ($key->judul) ?></h1></center>
                 </a>
-                <a href="<?php echo site_url('blog/Form_edit/'. $key->id); ?>" class="btn btn-sm btn-danger">edit</a>
+                <a href="<?php echo site_url('blog/edit/'. $key->id); ?>" class="btn btn-sm btn-danger">edit</a>
                 <a href='blog/delete/<?php echo $key->id ?>' class='btn btn-sm btn-danger'>Hapus</a>
               </td>
             </tr>
           </table>
         </div>
+        <?php endforeach; ?> 
         </div>
-      <?php endforeach ?>
+      
 
 
 
@@ -98,6 +116,11 @@
       </div><!-- /row -->
       </div> <!-- /container -->
   </div><!-- /white -->
+          <?php else : ?>
+        <p>Belum ada data bosque.</p>
+        <?php endif; ?>
+            </div>
+
   
   
   <!-- +++++ Footer Section +++++ -->
